backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings
httpx==0.27.0
pytest==8.1.1
black==24.4.0
isort==5.13.2
sqlalchemy[asyncio]
asyncpg
kiteconnect
passlib[bcrypt]
bcrypt>=4.0.1,<5.0
redis==5.0.1
websocket-client
